only the user making the change runs the hooks
Added a guard around all of the item and character update hooks so that
only the user making the change will actually run the update hooks.

If this is not in place, every logged in user that does not have
permission to modify the changed user will receive a permissions toast
error.